Reorganized some functions.
